pyproject.toml
[tool.poetry]
name = "potiguar-api"
version = "0.1.0"
description = ""
authors = ["Maxson Almeida <maxsonferovante@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
pyjwt = "^2.9.0"
pynest-api = "^0.3.1"
beanie = "1.20.0"
pytest = "^8.3.3"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
